Reduce kubelet LOC: extract cadvisor
